requirements.txt
pandas==2.3.0
python-dotenv==1.1.0
openai==1.88.0
Jinja2==3.1.6
frontmatter==3.0.8

# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Additional useful packages
requests==2.31.0
python-multipart==0.0.6